Single child table department approver
