+++ erpnext/hr/doctype/department/department.js
@@ -5,7 +5,7 @@ frappe.ui.form.on('Department', {
 	onload: function(frm) {
 		frm.set_query("leave_approver", "leave_approvers", function(doc) {
 			return {
-				query:"erpnext.hr.doctype.employee_leave_approver.employee_leave_approver.get_approvers",
+				query:"erpnext.hr.doctype.department_approver.department_approver.get_department_approvers",
 				filters:{
 					user: doc.user_id
 				}
@@ -13,7 +13,7 @@ frappe.ui.form.on('Department', {
 		});
 		frm.set_query("expense_approver", "expense_approvers", function(doc) {
 			return {
-				query:"erpnext.hr.doctype.employee_leave_approver.employee_leave_approver.get_approvers",
+				query:"erpnext.hr.doctype.department_approver.department_approver.get_department_approvers",
 				filters:{
 					user: doc.user_id
 				}
