+++ erpnext/hr/doctype/department/department.json
@@ -175,9 +175,8 @@
    "bold": 0, 
    "collapsible": 0, 
    "columns": 0, 
-   "description": "The first Leave Approver in the list will be set as the default Leave Approver.", 
-   "fieldname": "leave_approvers", 
-   "fieldtype": "Table", 
+   "fieldname": "leave_section", 
+   "fieldtype": "Section Break", 
    "hidden": 0, 
    "ignore_user_permissions": 0, 
    "ignore_xss_filter": 0, 
@@ -188,7 +187,70 @@
    "label": "Leave Approvers", 
    "length": 0, 
    "no_copy": 0, 
-   "options": "Employee Leave Approver", 
+   "permlevel": 0, 
+   "precision": "", 
+   "print_hide": 0, 
+   "print_hide_if_no_value": 0, 
+   "read_only": 0, 
+   "remember_last_selected_value": 0, 
+   "report_hide": 0, 
+   "reqd": 0, 
+   "search_index": 0, 
+   "set_only_once": 0, 
+   "translatable": 0, 
+   "unique": 0
+  }, 
+  {
+   "allow_bulk_edit": 0, 
+   "allow_on_submit": 0, 
+   "bold": 0, 
+   "collapsible": 0, 
+   "columns": 0, 
+   "description": "The first Leave Approver in the list will be set as the default Leave Approver.", 
+   "fieldname": "leave_approver", 
+   "fieldtype": "Table", 
+   "hidden": 0, 
+   "ignore_user_permissions": 0, 
+   "ignore_xss_filter": 0, 
+   "in_filter": 0, 
+   "in_global_search": 0, 
+   "in_list_view": 0, 
+   "in_standard_filter": 0, 
+   "label": "Leave Approver", 
+   "length": 0, 
+   "no_copy": 0, 
+   "options": "Department Approver", 
+   "permlevel": 0, 
+   "precision": "", 
+   "print_hide": 0, 
+   "print_hide_if_no_value": 0, 
+   "read_only": 0, 
+   "remember_last_selected_value": 0, 
+   "report_hide": 0, 
+   "reqd": 0, 
+   "search_index": 0, 
+   "set_only_once": 0, 
+   "translatable": 0, 
+   "unique": 0
+  }, 
+  {
+   "allow_bulk_edit": 0, 
+   "allow_on_submit": 0, 
+   "bold": 0, 
+   "collapsible": 0, 
+   "columns": 0, 
+   "fieldname": "expense_section", 
+   "fieldtype": "Section Break", 
+   "hidden": 0, 
+   "ignore_user_permissions": 0, 
+   "ignore_xss_filter": 0, 
+   "in_filter": 0, 
+   "in_global_search": 0, 
+   "in_list_view": 0, 
+   "in_standard_filter": 0, 
+   "label": "Expense Approvers", 
+   "length": 0, 
+   "no_copy": 0, 
    "permlevel": 0, 
    "precision": "", 
    "print_hide": 0, 
@@ -221,7 +283,7 @@
    "label": "Expense Approver", 
    "length": 0, 
    "no_copy": 0, 
-   "options": "Employee Expense Approver", 
+   "options": "Department Approver", 
    "permlevel": 0, 
    "precision": "", 
    "print_hide": 0, 
@@ -340,7 +402,7 @@
  "issingle": 0, 
  "istable": 0, 
  "max_attachments": 0, 
- "modified": "2018-04-13 15:53:35.792903", 
+ "modified": "2018-04-16 16:46:02.759012", 
  "modified_by": "Administrator", 
  "module": "HR", 
  "name": "Department",
