+++ erpnext/hr/doctype/department_approver/department_approver.json
@@ -1,63 +1,76 @@
 {
  "allow_copy": 0, 
+ "allow_guest_to_view": 0, 
  "allow_import": 0, 
  "allow_rename": 0, 
  "autoname": "hash", 
  "beta": 0, 
- "creation": "2013-04-12 06:56:15", 
+ "creation": "2018-04-08 16:31:02.433252", 
  "custom": 0, 
- "description": "Users who can approve a specific employee's leave applications", 
+ "description": "", 
  "docstatus": 0, 
  "doctype": "DocType", 
+ "document_type": "", 
  "editable_grid": 1, 
+ "engine": "InnoDB", 
  "fields": [
   {
+   "allow_bulk_edit": 0, 
    "allow_on_submit": 0, 
    "bold": 0, 
    "collapsible": 0, 
-   "fieldname": "leave_approver", 
+   "columns": 0, 
+   "fieldname": "approver", 
    "fieldtype": "Link", 
    "hidden": 0, 
    "ignore_user_permissions": 1, 
    "ignore_xss_filter": 0, 
    "in_filter": 0, 
+   "in_global_search": 0, 
    "in_list_view": 1, 
-   "label": "Leave Approver", 
+   "in_standard_filter": 0, 
+   "label": "Approver", 
    "length": 0, 
    "no_copy": 0, 
    "options": "User", 
    "permlevel": 0, 
+   "precision": "", 
    "print_hide": 1, 
    "print_hide_if_no_value": 0, 
    "read_only": 0, 
+   "remember_last_selected_value": 0, 
    "report_hide": 0, 
    "reqd": 1, 
    "search_index": 0, 
    "set_only_once": 0, 
+   "translatable": 0, 
    "unique": 0, 
    "width": "200"
   }
  ], 
+ "has_web_view": 0, 
  "hide_heading": 0, 
  "hide_toolbar": 0, 
- "idx": 1, 
+ "idx": 0, 
  "image_view": 0, 
  "in_create": 0, 
- "in_dialog": 0, 
  "is_submittable": 0, 
  "issingle": 0, 
  "istable": 1, 
  "max_attachments": 0, 
- "modified": "2016-07-11 03:28:00.270787", 
+ "modified": "2018-04-17 11:05:46.294340", 
  "modified_by": "Administrator", 
  "module": "HR", 
- "name": "Employee Leave Approver", 
+ "name": "Department Approver", 
+ "name_case": "", 
  "owner": "Administrator", 
  "permissions": [], 
  "quick_entry": 0, 
  "read_only": 0, 
  "read_only_onload": 0, 
+ "show_name_in_global_search": 0, 
  "sort_field": "modified", 
  "sort_order": "DESC", 
+ "track_changes": 0, 
  "track_seen": 0
 }
