--- erpnext/hr/doctype/employee_expense_approver/employee_expense_approver.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2018, Frappe Technologies Pvt. Ltd. and contributors
-# For license information, please see license.txt
-
-from __future__ import unicode_literals
-from frappe.model.document import Document
-
-class EmployeeExpenseApprover(Document):
-	pass
