+++ erpnext/hr/doctype/expense_claim/expense_claim.js
@@ -158,7 +158,7 @@ frappe.ui.form.on("Expense Claim", {
 	onload: function(frm) {
 		frm.set_query("expense_approver", function() {
 			return {
-				query: "erpnext.hr.doctype.leave_application.leave_application.get_approvers",
+				query: "erpnext.hr.doctype.department_approver.department_approver.get_approvers",
 				filters: {
 					employee: frm.doc.employee,
 					doctype: frm.doc.doctype
